refactor: ContentServer sync removed unused code and disconnect after sync
