+++ server/main.js
@@ -3,7 +3,6 @@ import '../imports/startup/server/services'
 import '../imports/startup/server/environment'
 import '../imports/startup/both/schema'
 import '../imports/startup/server/content'
-import '../imports/startup/server/sync'
 import '../imports/startup/server/accounts'
 import '../imports/startup/server/users'
 import '../imports/startup/server/teacherContexts'
